Modified and cleaned version of ICL protocol implementation
- fixed inheritance and duplicated code from base clases
- removed Daisy printer specific implementation due to higher level of implementation(will be separated in another project)
- fixed scope of some properties to 'protected' to be available in descendant classes
- remove incompatible license headers
- fixed graddle default file encoding to 'UTF-8'
